+++ services/web/server/src/simcore_service_webserver/projects/_projects_service.py
@@ -30,7 +30,6 @@ from models_library.api_schemas_dynamic_scheduler.dynamic_services import (
     DynamicServiceStop,
 )
 from models_library.api_schemas_webserver.projects import (
-    ProjectDocument,
     ProjectGet,
     ProjectPatch,
 )
@@ -39,8 +38,7 @@ from models_library.basic_types import KeyIDStr
 from models_library.errors import ErrorDict
 from models_library.groups import GroupID
 from models_library.products import ProductName
-from models_library.projects import Project, ProjectID, ProjectTemplateType
-from models_library.projects import ProjectType as ProjectTypeAPI
+from models_library.projects import Project, ProjectID
 from models_library.projects_access import Owner
 from models_library.projects_nodes import Node, NodeState, PartialNode
 from models_library.projects_nodes_io import NodeID, NodeIDStr, PortLink
@@ -85,10 +83,8 @@ from servicelib.rabbitmq.rpc_interfaces.dynamic_scheduler.errors import (
     ServiceWasNotFoundError,
 )
 from servicelib.redis import (
-    PROJECT_DB_UPDATE_REDIS_LOCK_KEY,
     exclusive,
     get_project_locked_state,
-    increment_and_return_project_document_version,
     is_project_locked,
     with_project_locked,
 )
@@ -107,10 +103,7 @@ from ..director_v2 import director_v2_service
 from ..dynamic_scheduler import api as dynamic_scheduler_service
 from ..products import products_web
 from ..rabbitmq import get_rabbitmq_rpc_client
-from ..redis import (
-    get_redis_document_manager_client_sdk,
-    get_redis_lock_manager_client_sdk,
-)
+from ..redis import get_redis_lock_manager_client_sdk
 from ..resource_manager.models import UserSession
 from ..resource_manager.user_sessions import (
     PROJECT_ID_KEY,
@@ -146,6 +139,7 @@ from ._access_rights_service import (
     has_user_project_access_rights,
 )
 from ._nodes_utils import set_reservation_same_as_limit, validate_new_service_resources
+from ._project_document_utils import build_project_document_and_increment_version
 from ._projects_repository_legacy import APP_PROJECT_DBAPI, ProjectDBAPI
 from ._projects_repository_legacy_utils import PermissionStr
 from ._socketio_service import notify_project_document_updated
@@ -209,46 +203,8 @@ async def patch_project_and_notify_users(
         new_partial_project_data=patch_project_data,
     )
 
-    @exclusive(
-        get_redis_lock_manager_client_sdk(app),
-        lock_key=PROJECT_DB_UPDATE_REDIS_LOCK_KEY.format(project_uuid),
-        blocking=True,
-        blocking_timeout=None,  # NOTE: this is a blocking call, a timeout has undefined effects
-    )
-    async def _build_project_document_and_increment_version() -> (
-        tuple[ProjectDocument, int]
-    ):
-        """This function is protected because
-        - the project document and its version must be kept in sync
-        """
-        project_with_workbench = await _projects_repository.get_project_with_workbench(
-            app=app, project_uuid=project_uuid
-        )
-        project_document = ProjectDocument(
-            uuid=project_with_workbench.uuid,
-            workspace_id=project_with_workbench.workspace_id,
-            name=project_with_workbench.name,
-            description=project_with_workbench.description,
-            thumbnail=project_with_workbench.thumbnail,
-            last_change_date=project_with_workbench.last_change_date,
-            classifiers=project_with_workbench.classifiers,
-            dev=project_with_workbench.dev,
-            quality=project_with_workbench.quality,
-            workbench=project_with_workbench.workbench,
-            ui=project_with_workbench.ui,
-            type=cast(ProjectTypeAPI, project_with_workbench.type),
-            template_type=cast(
-                ProjectTemplateType, project_with_workbench.template_type
-            ),
-        )
-        redis_client_sdk = get_redis_document_manager_client_sdk(app)
-        document_version = await increment_and_return_project_document_version(
-            redis_client=redis_client_sdk, project_uuid=project_uuid
-        )
-        return project_document, document_version
-
     project_document, document_version = (
-        await _build_project_document_and_increment_version()
+        await build_project_document_and_increment_version(app, project_uuid)
     )
     await notify_project_document_updated(
         app=app,
